Create new hierarhy for ISV-repos and link acme to the official one
Previously it was expected that users will link the repository using
snactor. But possibly many solutions could be created just for purpose
of inplace upgrade between specific systems (e.g. RHEL 7 -> 8) and
will not be valid for upgrade (RHEL 8 -> 9).

To make it clear, let's use the el7toel8 directory.

As the ACME repository clearly is now just for inplace upgrade from
RHEL 7 to RHEL 8, link it directly to the official
system_upgrade_el7toel8 repository.